requirements.txt
# 核心依赖
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM API客户端
openai>=1.12.0
anthropic>=0.18.0

# 配置和日志
pyyaml>=6.0
python-dotenv>=1.0.0

# 邮件发送
secure-smtplib>=0.1.1

# 调度
schedule>=1.2.0
pytz>=2023.3

# 数据处理
python-dateutil>=2.8.2

# HTTP客户端增强
urllib3>=2.0.0
certifi>=2023.7.22

# HTML解析
html5lib>=1.1

